+++ Sources/TuistKit/Services/FocusService.swift
@@ -9,13 +9,15 @@ import TuistLoader
 import TuistSupport
 
 protocol FocusServiceProjectGeneratorFactorying {
-    func generator(sources: Set<String>, xcframeworks: Bool) -> ProjectGenerating
+    func generator(sources: Set<String>, xcframeworks: Bool, ignoreCache: Bool) -> ProjectGenerating
 }
 
 final class FocusServiceProjectGeneratorFactory: FocusServiceProjectGeneratorFactorying {
-    func generator(sources: Set<String>, xcframeworks: Bool) -> ProjectGenerating {
+    func generator(sources: Set<String>, xcframeworks: Bool, ignoreCache: Bool) -> ProjectGenerating {
         let cacheOutputType: CacheOutputType = xcframeworks ? .xcframework : .framework
-        let cacheConfig = CacheConfig.withCaching(cacheOutputType: cacheOutputType)
+        let cacheConfig: CacheConfig = ignoreCache
+            ? .withoutCaching()
+            : .withCaching(cacheOutputType: cacheOutputType)
         return ProjectGenerator(graphMapperProvider: GraphMapperProvider(cacheConfig: cacheConfig, sources: sources))
     }
 }
@@ -51,12 +53,14 @@ final class FocusService {
         self.projectGeneratorFactory = projectGeneratorFactory
     }
 
-    func run(path: String?, sources: Set<String>, noOpen: Bool, xcframeworks: Bool) throws {
+    func run(path: String?, sources: Set<String>, noOpen: Bool, xcframeworks: Bool, ignoreCache: Bool) throws {
         let path = self.path(path)
         if isWorkspace(path: path) {
             throw FocusServiceError.cacheWorkspaceNonSupported
         }
-        let generator = projectGeneratorFactory.generator(sources: sources, xcframeworks: xcframeworks)
+        let generator = projectGeneratorFactory.generator(sources: sources,
+                                                          xcframeworks: xcframeworks,
+                                                          ignoreCache: ignoreCache)
         let workspacePath = try generator.generate(path: path, projectOnly: false)
         if !noOpen {
             try opener.open(path: workspacePath)