+++ Tests/TuistKitTests/Services/FocusServiceTests.swift
@@ -9,18 +9,20 @@ import XCTest
 @testable import TuistLoaderTesting
 @testable import TuistSupportTesting
 
+private typealias GeneratorParameters = (sources: Set<String>, xcframeworks: Bool, ignoreCache: Bool)
+
 final class MockFocusServiceProjectGeneratorFactory: FocusServiceProjectGeneratorFactorying {
     var invokedGenerator = false
     var invokedGeneratorCount = 0
-    var invokedGeneratorParameters: (sources: Set<String>, xcframeworks: Bool)?
-    var invokedGeneratorParametersList = [(sources: Set<String>, xcframeworks: Bool)]()
+    fileprivate var invokedGeneratorParameters: GeneratorParameters?
+    fileprivate var invokedGeneratorParametersList = [GeneratorParameters]()
     var stubbedGeneratorResult: ProjectGenerating!
 
-    func generator(sources: Set<String>, xcframeworks: Bool) -> ProjectGenerating {
+    func generator(sources: Set<String>, xcframeworks: Bool, ignoreCache: Bool) -> ProjectGenerating {
         invokedGenerator = true
         invokedGeneratorCount += 1
-        invokedGeneratorParameters = (sources, xcframeworks)
-        invokedGeneratorParametersList.append((sources, xcframeworks))
+        invokedGeneratorParameters = (sources, xcframeworks, ignoreCache)
+        invokedGeneratorParametersList.append((sources, xcframeworks, ignoreCache))
         return stubbedGeneratorResult
     }
 }
@@ -64,7 +66,7 @@ final class FocusServiceTests: TuistUnitTestCase {
             throw error
         }
 
-        XCTAssertThrowsError(try subject.run(path: nil, sources: Set(), noOpen: true, xcframeworks: false)) {
+        XCTAssertThrowsError(try subject.run(path: nil, sources: Set(), noOpen: true, xcframeworks: false, ignoreCache: false)) {
             XCTAssertEqual($0 as NSError?, error)
         }
     }
@@ -76,7 +78,7 @@ final class FocusServiceTests: TuistUnitTestCase {
             workspacePath
         }
 
-        try subject.run(path: nil, sources: Set(), noOpen: false, xcframeworks: false)
+        try subject.run(path: nil, sources: Set(), noOpen: false, xcframeworks: false, ignoreCache: false)
 
         XCTAssertEqual(opener.openArgs.last?.0, workspacePath.pathString)
     }